authorization implemented

+++ src/api/notes/handler.js
@@ -14,8 +14,14 @@ class NotesHandler {
     try {
       this._validator.validateNotePayload(request.payload);
       const { title = "untitled", body, tags } = request.payload;
+      const { id: credentialId } = request.auth.credentials;
 
-      const noteId = await this._services.addNote({ title, body, tags });
+      const noteId = await this._services.addNote({
+        title,
+        body,
+        tags,
+        owner: credentialId,
+      });
 
       const response = h.response({
         status: "success",
@@ -45,8 +51,9 @@ class NotesHandler {
       return response;
     }
   }
-  async getNotesHandler() {
-    const notes = await this._services.getNotes();
+  async getNotesHandler(request) {
+    const { id: credentialId } = request.auth.credentials;
+    const notes = await this._services.getNotes(credentialId);
     return {
       status: "success",
       data: {
@@ -57,6 +64,9 @@ class NotesHandler {
   async getNoteByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+
+      await this._services.verifyNoteOwner(id, credentialId);
       const note = await this._services.getNoteById(id);
       return {
         status: "success",
@@ -87,6 +97,8 @@ class NotesHandler {
     try {
       const { id } = request.params;
       this._validator.validateNotePayload(request.payload);
+      const { id: credentialId } = request.auth.credentials;
+      await this._services.verifyNoteOwner(id, credentialId);
       await this._services.editNoteById(id, request.payload);
 
       return {
@@ -116,6 +128,8 @@ class NotesHandler {
   async deleteNoteByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+      await this._services.verifyNoteOwner(id, credentialId);
       await this._services.deleteNoteById(id);
       return {
         status: "success",
